fix: explicitly mark unreachable code to prevent GCC warnings
Currently, GCC warns about code that switches on an enum, but despite
handling all enumerations, will not return a value in the default case.

    thrift/compiler/sema/check_map_keys.cc: In function 'bool apache::thrift::compiler::{anonymous}::lt_value(const apache::thrift::compiler::t_const_value*, const apache::thrift::compiler::t_const_value*)':
    thrift/compiler/sema/check_map_keys.cc:167:1: warning: control reaches end of non-void function [-Wreturn-type]
      167 | }
          | ^

This change adds a new header `thrift/compiler/portability.h` that
defines an `assume_unreachable()` function that can be used to mark
these locations as unreachable explicitly.

I'm not sure why the compiler doesn't use folly outside of the tests,
but using `folly::assume_unreachable()` would obviously work as well,
and it's already being used elsewhere in fbthrift.